refactor: extract ResponsiveModal as shared component
- Create ResponsiveModal component for cross-device UI optimization
- Extract PC (Dialog) and SP (Bottom Sheet) logic into reusable component
- Simplify CreateDraftModal to focus on domain logic only
- Implement proper separation of UI and domain concerns
- Add comprehensive documentation updates for modal system
- Maintain full functionality while improving maintainability

Benefits:
- DRY principle: eliminates UI code duplication
- Scalability: new modals can reuse responsive behavior
- Clean architecture: features/ focuses on business logic
- Consistency: unified UX across all future modals
